Move preprint PDF self-uri code, and modify tag after publication
