+++ provider/preprint.py
@@ -405,3 +405,62 @@ def get_preprint_pdf_url(endpoint_url, caller_name, user_agent=None):
             % (caller_name, response.status_code, endpoint_url)
         )
     return pdf_url
+
+
+def generate_new_pdf_href(article_id, version, content_subfolder):
+    "generate a new name for a preprint PDF"
+    new_pdf_file_name = PREPRINT_PDF_FILE_NAME_PATTERN.format(
+        article_id=utils.pad_msid(article_id), version=version
+    )
+    new_pdf_href = "/".join(
+        [part for part in [content_subfolder, new_pdf_file_name] if part]
+    )
+    return new_pdf_href
+
+
+def clear_pdf_self_uri(xml_root):
+    "remove self-uri tag if its content-type is pdf"
+    article_meta_tag = xml_root.find(".//front/article-meta")
+    for self_uri_tag in article_meta_tag.findall('self-uri[@content-type="pdf"]'):
+        article_meta_tag.remove(self_uri_tag)
+
+
+def set_pdf_self_uri_tag(xml_root, pdf_file_name, identifier):
+    "modify self-uri XML tag"
+    # remove old self-uri tags
+    clear_pdf_self_uri(xml_root)
+
+    # determine where to insert self-uri tag
+    insert_index = 0
+    article_meta_tag = xml_root.find(".//front/article-meta")
+    for index, tag in enumerate(article_meta_tag.findall("*")):
+        if tag.tag in ["permissions"]:
+            insert_index = index + 1
+            break
+
+    # add pdf self-uri tag
+    self_uri_tag = Element("self-uri")
+    self_uri_tag.set("content-type", "pdf")
+    self_uri_tag.set("{http://www.w3.org/1999/xlink}href", pdf_file_name)
+    self_uri_tag.tail = "\n"
+    article_meta_tag.insert(insert_index, self_uri_tag)
+
+
+def set_pdf_self_uri(xml_file_path, pdf_file_name, identifier):
+    "set or add a self-uri tag to article XML for an article PDF file"
+    # Register namespaces
+    xmlio.register_xmlns()
+
+    # get the XML doctype
+    xml_root, doctype_dict, processing_instructions = xmlio.parse(
+        xml_file_path,
+        return_doctype_dict=True,
+        return_processing_instructions=True,
+    )
+
+    set_pdf_self_uri_tag(xml_root, pdf_file_name, identifier)
+
+    # write the XML root to disk
+    cleaner.write_xml_file(
+        xml_root, xml_file_path, identifier, doctype_dict, processing_instructions
+    )
